Rss refactored
