+++ front/client/components/CategoryInfo.vue
@@ -33,7 +33,7 @@
       class="ma-0 px-1"
       flat
       color="grey"
-      @click="openAddItemDialog"
+      @click="openDialog"
     >
       <v-icon size="14" class="mr-1" left>$vuetify.icons.plus</v-icon>
       Add new item
@@ -52,10 +52,6 @@
       v-model="isCategoryInfoEdit"
       :categoryId="categoryId"
     />
-    <add-item-dialog
-      v-model="isAddItemDialogOpen"
-      :categoryId="categoryId"
-    />
   </div>  
 </template>
 
@@ -64,17 +60,19 @@ import Vue from 'vue'
 import Component from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 import CategoryInfoEdit from 'client/components/CategoryInfoEdit.vue'
-import AddItemDialog from 'client/components/AddItemDialog.vue'
 import ALink from 'client/components/ALink.vue'
+import Confirm from 'client/helpers/ConfirmDecorator'
+import CategoryItemBtn from 'client/components/CategoryItemBtn.vue'
 
 @Component({
   components: {
     CategoryInfoEdit,
-    AddItemDialog,
     ALink,
+    CategoryItemBtn,
   }
 })
 export default class CategoryInfo extends Vue {
+  @Prop(Object) category!: object
   @Prop(String) categoryId!: string
   @Prop(String) categoryTitle!: string
   @Prop(String) categoryGroup!: string
@@ -87,8 +85,61 @@ export default class CategoryInfo extends Vue {
     this.isCategoryInfoEdit = true
   }
 
-  openAddItemDialog () {
-    this.isAddItemDialogOpen = true
+  openDialog () {
+    this.$emit('openDialog');
+  }
+
+  @Confirm({ text: 'delete this category' })
+  async deleteCategory () {
+    if (!this.category) {
+      return
+    }
+    await this.$store.dispatch('category/deleteCategory', this.categoryId)
+    this.$router.back()
   }
 }
 </script>
+
+<style scoped>
+.category-top {
+  display: flex;
+  align-items: center;
+  margin: 0 0 5px;
+}
+
+.category-top-data {
+  display: flex;
+  align-items: center;
+  flex: 1;
+}
+
+.category-top >>> i {
+  margin-right: 15px;
+  font-size: 18px;
+  color: #979797;
+  cursor: pointer;
+  transition: all ease-in-out 0.25s;
+}
+
+.category-top >>> i:hover {
+  color: #000;
+}
+
+.category-top-link {
+  font-size: 24px;
+  font-weight: 600;
+  text-decoration: none;
+  color: #979797;
+  cursor: pointer;
+  transition: all ease-in-out 0.25s;
+}
+
+.category-top-link:hover {
+  color: #000;
+}
+
+.category-top-group {
+  font-size: 24px;
+}
+</style>
+
