+++ front/config.json
@@ -1,3 +1,3 @@
 {
-  "apiUrl": "http://guide.aelve.com:4800/api"
+  "apiUrl": "http://localhost:4400"
 }
